+++ packages/ice/src/config.ts
@@ -122,7 +122,7 @@ const userConfig = [
   {
     name: 'ssr',
     validation: 'boolean',
-    defaultValue: true,
+    defaultValue: false,
   },
   {
     name: 'server',
@@ -358,7 +358,7 @@ const userConfig = [
 
 const cliOption = [
   {
-    name: 'platform',
+    name: 'target',
     commands: ['start', 'build'],
   },
   {