+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v3.0.6
+
+- [feat] support `import.meta.target`, `import.meta.renderer` and `import.meta.env`
+- [feat] support config `routes.injectInitialEntry` for memory router
+- [fix] cli option `--platform` change to `--target`
+- [fix] remove code when dead code is `ObjectProperties`
+- [fix] rebuild server entry when document changed
+- [fix] set ssr default value to `false`
+- [fix] server compiler options when bundle server entry
+- [fix] support define expression of basename
+
 ## v3.0.5
 
 - [fix] bump @swc/helpers version (0.4.12-> 0.4.14) for module cannot resolve of `@swc/helpers/src/_object_destructuring_empty.mjs`