+++ packages/plugin-rax-compat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.3
+
+- [refactor] Using rax-compat/jsx-runtime to support rax components
+
 ## 0.1.2
 
 - [fix] get css path by onLoad lifecycle.