+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @ice/runtime
 
+## v1.0.5
+
+- [feat] support template parse for dataLoader
+- [feat] enhance memory router
+
 ## v1.0.4
 
 - [chore] show more log when failed to load route module